refactor: CardDetail 관련 컴포넌트 CardDetailPage 하위로 이동
