+++ src/component/CardDetailPage/CardDetailHeader/CardDetailHeader.stories.tsx
