+++ src/component/CardDetailPage/CardDetailHeader/styled.tsx
