+++ src/component/CardDetailPage/CardDetailView/CardDetailView.stories.tsx
