+++ src/component/CardDetailPage/CardDetailView/index.tsx
