+++ src/component/CardDetailPage/CardDetailView/styled.tsx
