+++ src/component/CardDetailPage/index.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import St from "./styled";
-import CardDetailHeader from "../CardDetailHeader";
-import CardDetailView from "../CardDetailView";
-import CardDetailForm from "../CardDetailForm";
-
+import CardDetailHeader from "./CardDetailHeader";
+import CardDetailView from "./CardDetailView";
+import CardDetailForm from "./CardDetailForm";
 function CardDetailPage() {
   return (
     <St.Page>
